# docs/swagger.yaml
definitions:
  analytics.BatchRequest:
    properties:
      batch_id:
        type: string
      events:
        items:
          $ref: '#/definitions/analytics.EventInput'
        type: array
      sent_at:
        type: string
    type: object
  analytics.ConsentRequest:
    properties:
      consent:
        type: boolean
    type: object
  analytics.ConsentResponse:
    properties:
      consent:
        type: boolean
    type: object
  analytics.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  analytics.IngestResponse:
    properties:
      accepted:
        type: integer
      deduplicated:
        type: integer
    type: object
  auth.AuthRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  auth.PasswordChangeConfirmRequest:
    properties:
      changeToken:
        type: string
      password:
        type: string
      passwordConfirm:
        type: string
    type: object
  auth.PasswordChangeVerifyRequest:
    properties:
      code:
        type: string
    type: object
  auth.PasswordChangeVerifyResponse:
    properties:
      changeToken:
        type: string
    type: object
  auth.PasswordResetConfirmRequest:
    properties:
      password:
        type: string
      passwordConfirm:
        type: string
      resetToken:
        type: string
    type: object
  auth.PasswordResetRequest:
    properties:
      email:
        type: string
    type: object
  auth.PasswordResetVerifyRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  auth.PasswordResetVerifyResponse:
    properties:
      resetToken:
        type: string
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
      refreshToken:
        type: string
    type: object
  auth.TokenPair:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  auth.UpdateProfileRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
    type: object
  auth.UserProfileResponse:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
    type: object
...
  vitamins.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API для регистрации и авторизации
  title: Vitamins API
  version: "1.0"
paths:
  /admin/analytics/export:
    get:
      parameters:
      - description: RFC3339 from
        in: query
        name: from
        type: string
      - description: RFC3339 to
        in: query
        name: to
        type: string
      - description: event_name
        in: query
        name: event
        type: string
      - description: user_id
        in: query
        name: user_id
        type: integer
      - description: csv|jsonl
        in: query
        name: format
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - text/csv
      - application/json
      responses:
        "200":
          description: CSV/JSONL
          schema:
            type: string
        "401":
          description: ADMIN_REQUIRED
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
      summary: Экспорт аналитики
      tags:
      - admin
  /analytics/consent:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.ConsentResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить согласие на аналитику
      tags:
      - analytics
    post:
      consumes:
      - application/json
      parameters:
      - description: Согласие
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analytics.ConsentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.ConsentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Установить согласие на аналитику
      tags:
      - analytics
  /analytics/events:
    post:
      consumes:
      - application/json
      description: Принимает батч аналитических событий
      parameters:
      - description: Batch событий
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/analytics.BatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/analytics.IngestResponse'
        "400":
          description: EMPTY_BATCH, BATCH_TOO_LARGE, INVALID_EVENT_ID, INVALID_OCCURRED_AT,
            INVALID_EVENT_NAME, INVALID_SESSION_ID, INVALID_ANONYMOUS_ID, ANONYMOUS_ID_REQUIRED
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "401":
          description: INVALID_TOKEN
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "403":
          description: CONSENT_REQUIRED
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/analytics.ErrorResponse'
      summary: Прием аналитических событий
      tags:
      - analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет учетные данные и возвращает новую пару токенов
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: EMAIL_AND_PASSWORD_REQUIRED, EMAIL_REQUIRED, INVALID_EMAIL_FORMAT,
            PASSWORD_REQUIRED, INVALID_PASSWORD_FORMAT
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: INVALID_CREDENTIALS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/password/reset/confirm:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по resetToken
      parameters:
      - description: Reset token и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: PASSWORD_REQUIRED, INVALID_PASSWORD_FORMAT, PASSWORD_CONFIRMATION_MISMATCH,
            RESET_TOKEN_REQUIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: RESET_SESSION_INVALID, RESET_SESSION_EXPIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Смена пароля
      tags:
      - auth
  /auth/password/reset/request:
    post:
      consumes:
      - application/json
      description: Отправляет код на e-mail через SMTP
      parameters:
      - description: E-mail
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: EMAIL_REQUIRED, INVALID_EMAIL_FORMAT
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "429":
          description: TOO_MANY_REQUESTS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Запрос кода восстановления пароля
      tags:
      - auth
  /auth/password/reset/verify:
    post:
      consumes:
      - application/json
      description: Проверяет код и возвращает resetToken
      parameters:
      - description: E-mail и код
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordResetVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.PasswordResetVerifyResponse'
        "400":
          description: EMAIL_REQUIRED, INVALID_EMAIL_FORMAT, RESET_CODE_REQUIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: RESET_CODE_INVALID, RESET_CODE_EXPIRED, RESET_CODE_TOO_MANY_ATTEMPTS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Проверка кода восстановления
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access token по refresh token из Authorization Bearer
        и/или тела запроса
      parameters:
      - description: Refresh token (refreshToken или refresh_token)
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      - description: Bearer <refreshToken>
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: BAD_REQUEST
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: INVALID_REFRESH_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает учетную запись и возвращает пару access/refresh токенов
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        "400":
          description: EMAIL_AND_PASSWORD_REQUIRED, EMAIL_REQUIRED, INVALID_EMAIL_FORMAT,
            PASSWORD_REQUIRED, INVALID_PASSWORD_FORMAT
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: EMAIL_ALREADY_EXISTS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /users/me:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о текущем пользователе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserProfileResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Профиль пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Обновляет имя, фамилию и/или e-mail текущего пользователя
      parameters:
      - description: Данные профиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.UserProfileResponse'
        "400":
          description: NO_FIELDS_TO_UPDATE, EMAIL_REQUIRED, INVALID_EMAIL_FORMAT
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: EMAIL_ALREADY_EXISTS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновление профиля
      tags:
      - users
    delete:
      consumes:
      - application/json
      description: Удаляет текущий аккаунт и связанные пользовательские данные
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление аккаунта
      tags:
      - users
  /users/me/password/change/confirm:
    post:
      consumes:
      - application/json
      description: Устанавливает новый пароль по changeToken
      parameters:
      - description: Change token и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordChangeConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: PASSWORD_REQUIRED, INVALID_PASSWORD_FORMAT, PASSWORD_CONFIRMATION_MISMATCH,
            CHANGE_TOKEN_REQUIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: CHANGE_SESSION_INVALID, CHANGE_SESSION_EXPIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: REDIS_NOT_CONFIGURED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Смена пароля
      tags:
      - users
  /users/me/password/change/request:
    post:
      consumes:
      - application/json
      description: Отправляет код на e-mail текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "429":
          description: TOO_MANY_REQUESTS
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: MAILER_NOT_CONFIGURED, REDIS_NOT_CONFIGURED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Запрос кода смены пароля
      tags:
      - users
  /users/me/password/change/verify:
    post:
      consumes:
      - application/json
      description: Проверяет код и возвращает changeToken
      parameters:
      - description: Код из письма
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.PasswordChangeVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.PasswordChangeVerifyResponse'
        "400":
          description: CHANGE_CODE_REQUIRED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: CHANGE_CODE_INVALID, CHANGE_CODE_TOO_MANY_ATTEMPTS, AUTH_REQUIRED,
            INVALID_TOKEN
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "404":
          description: USER_NOT_FOUND
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: REDIS_NOT_CONFIGURED
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Проверка кода смены пароля
      tags:
      - users
  /vitamins/catalog:
    get:
      consumes:
      - application/json
      description: Возвращает список витаминов из справочника
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vitamins.CatalogItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      summary: Каталог витаминов
      tags:
      - vitamins
  /vitamins/reminders:
    get:
      consumes:
      - application/json
      description: Возвращает список напоминаний пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/vitamins.ReminderResponse'
            type: array
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список напоминаний
      tags:
      - vitamins
    post:
      consumes:
      - application/json
      description: Создает пользовательский витамин с курсом, расписанием и настройками
        уведомления
      parameters:
      - description: Данные напоминания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vitamins.CreateReminderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: NAME_REQUIRED, INVALID_FORM, INVALID_CONDITION, INVALID_DAYS,
            INVALID_TIMES, START_DATE_REQUIRED, INVALID_DATE_FORMAT, INVALID_COURSE_DURATION,
            TIMEZONE_REQUIRED
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: CATALOG_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать напоминание о витамине
      tags:
      - vitamins
  /vitamins/reminders/{id}:
    delete:
      consumes:
      - application/json
      description: Мягко удаляет напоминание (is_active=false)
      parameters:
      - description: ID напоминания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: REMINDER_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить напоминание
      tags:
      - vitamins
    get:
      consumes:
      - application/json
      description: Возвращает подробную информацию о напоминании
      parameters:
      - description: ID напоминания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: REMINDER_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Детали напоминания
      tags:
      - vitamins
    patch:
      consumes:
      - application/json
      description: Обновляет дозу, примечание, условие, расписание, курс и настройки
        уведомлений
      parameters:
      - description: ID напоминания
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/vitamins.UpdateReminderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: INVALID_ID, NO_FIELDS_TO_UPDATE, NAME_REQUIRED, INVALID_FORM,
            INVALID_CONDITION, INVALID_DAYS, INVALID_TIMES, START_DATE_REQUIRED, INVALID_DATE_FORMAT,
            INVALID_COURSE_DURATION, TIMEZONE_REQUIRED
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: REMINDER_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить напоминание
      tags:
      - vitamins
  /vitamins/reminders/{id}/disable:
    post:
      consumes:
      - application/json
      description: Устанавливает is_active=false
      parameters:
      - description: ID напоминания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: REMINDER_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Отключить напоминание
      tags:
      - vitamins
  /vitamins/reminders/{id}/enable:
    post:
      consumes:
      - application/json
      description: Устанавливает is_active=true
      parameters:
      - description: ID напоминания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vitamins.ReminderResponse'
        "400":
          description: INVALID_ID
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "401":
          description: AUTH_REQUIRED, INVALID_TOKEN
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "404":
          description: REMINDER_NOT_FOUND
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/vitamins.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Включить напоминание
      tags:
      - vitamins
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"